Simplify logic regarding missing parameter types
When we're typing the result of eta-expansion and fail
because of missing parameter types, we try again.
Whether we're in that case doesn't change between arguments,
so hoist that out of the loop.

Eventually, we should get rid of the retry and just immediately
type check the un-eta-expanded method call -- if possible...