TODO: test `match` without default case
For example, removing the `case _ =>` from the test case:

```scala
  def a(x1: Any, x2: Any): Int = (x1, x2) match {
    case (s1: String, s2: String) => return 1
    case (i1: Int, s2)            => return 2
    case (s1, s2: String)         => return 3
  }
```
Currently, this automatically adds the following default case:
```scala
case tup => throw new MatchError(tup)
```
which, AFAIU, would prevent this optimization from applying because it captures `tup`?

Instead, the above could generate a different synthetic default case:
```scala
case _ => throw new MatchError((x1, x2))
```

Since the pair was just created to do the `match`, this is semantically
equivalent, and would allow the optimization to apply.

Instead, this can always be done on a case-by-case basis: every case that
captures the just-constructed tuple, of the form:

```scala
case tup if G => body
```
would be rewritten as:
```scala
case _ if G => val tup = (x1, x2); body
```

thereby avoiding the capture of the tuple instance created for the match,
without changing the semantics. And then the optimization would apply. For
those branches that captured `tup`, of course they will still do the
allocation. But the branches that *don't* won't need the allocation, and these
are supposed to be the good paths.

(This is complicated if `G` refers to `tup`, of course, in which case that
might not be possible in the general case. But usually it doesn't.)

This is precisely the only problem we have with tuple-matches in the Scala.js
optimizer. If there is a default case, our optimizer is already able to do all
that this PR brings, simply as a consequence of stack allocation. But the
synthetic default case that puts `tup` in a `MatchError` kills us so far.