Allow mixed builds in ant
The JFunction classes depend on the FunctionN traits, so the java
compiler needs the Scala library on the classpath.

At the same time, while compiling the Scala library, the symbols for
JFunction classes need to be available to emit indy-lambda closures.
Therefore we pass the JFunctions as Java sources while compiling the
Scala library.