SI-2897, SI-6231 method-local traits capturing variables
Instead of passing the ObjectRef to the impl class's ctor,
which would store it in a field in the impl class,
which is then mixed in to the subclass, reworking the scheme.

The new scheme doesn't introduce fields/ctor args
for captured variables in traits, only a getter for it.

The getter is implemented in subclasses that mix in
the trait, which have the captured variable in scope.