Improve repl error positions, tree gen
Parse user input as-is, and use that as content of
unit's source file content. The unit's body has the
trees resulting from wrapping and rewriting the user's
input to implement the repl's semantics (store the
result in a val etc).

This allows us to be more precise in our error messages,
as positions map 1-1 to user input (as suggested by retronym).

Goal: no string-based code manipulation. No there yet.