More precise prototype for args of overloaded method
Specifically, for arguments of function-ish types to overloaded methods,
we try to preserve the expected function parameter types captured
in the types expected for their arguments, so that we can infer the
parameter types of functions passed as arguments.

Imagine an overloaded method `m_i`, with `N` overloads `i = 1..N`,
and an expected argument type at index `j`, `a_ij`:

```
def m_1(... a_1j, ...)
..
def m_N(... a_Nj, ...)
```

The expected type for the argument at index `j` will be more
precise than the usual `WildcardType` (`?`), if all types `a_1j..a_Nj`
are function-ish types that denote the same parameter types `p1..pM`.

A "function-ish" type is a `FunctionN[p1,...,pM]`, including `PartialFunction`,
or the equivalent SAM type.

The non-wildcard expected type will be
  - `PartialFunction[p1, ?]`, if an `a_ij` expects a partial function;
  - else, if there is a subclass of `FunctionM` among the `a_ij`,
    it is `FunctionM[p1, pM, ?]`;
  - otherwise, if all `a_ij` are equivalent under =:=, it is `a_ij`.

In each case, any type parameter not already resolved by overloading,
or the outer context, it approximated by `?`.

This allows better typing for an overloaded higher-order method
such as `map` (prevalent in the new collections / spark apis),
regardless of whether the actual argument is a function literal,
a method reference that will be eta-expanded (based on the expected type),
or a method value (`m _`).