co-evolve more eagerly
Given

```
object T { self: {type T = Int} =>
  type T
  def lift(x: Int): T = x
}

T.lift(1): Int
```

We want the last line to result in an error, but this
requires patching up the sig of `lift` as seen from `T`,
to replace the type alias `T` defined in the self type
by the abstract type symbol defined in `object T`.

The debuglog should contain a line like this:

```
co-evolve: <refinement>.type -> Test.this.T.type, type T : scala.this.Int -> type T
```

`<refinement>.type` is the self type, and `Test.this.T.type` is
the new prefix (since we're calling `lift` on `T`). The old symbol
is the alias to `Int`, whereas the correct symbol is the abstract type.