Subclassing on traits does not imply suffixing for their linearisations
this is true for non-trait classes: the linearisation of a class C contains
as segments the linearisations of its (non-trait) super classes, so we can
just treat those at the lowest point in the hierarchy

this optimization breaks down when considering traits and their linearisations,
as shown in the tests

Part 2: consider more symbol pairs in refchecks

we used to exclude/optimize a bit too much, on the erroneous assumption
that we would consider an overriding pair redundantly. We can only
eliminate pairs where we are sure they will occur in the same (linearisation) order,
which means trait-defined methods are not eligible.

note that Java-defined members implicitly override matching members
(Java doesn't have Scala's `override` keyword).

Now we see more pairs, had to refine a few mechanisms that defend
against double error reporting.