Bring Stepper.to method in line with collections
In java8-compat, Stepper has a `.accumulate` extension method. Now
that Accumulators are Scala collections, we can use `.to(Accumulator)`
instead.

Unlike Accumulators, I didn't integrate steppers into the collections
hierarchy. They are really intended to expose collections for the use
in Java streams, but not intended to be used as a colleciton. The
`.iterator` method is available to get a Scala iterator.

TODO: add deprecated `.accumulate` extension methods (for scala
collections, arrays, steppers, java streams)